requirements.txt
# Workorder System - Python Dependencies
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# Image processing
filelock==3.13.1
Pillow==10.2.0

# Environment management
python-dotenv==1.0.1

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0